+++ src/app/module/admin/forms/customer-info/customer-search/customer-search.component.ts
@@ -3,6 +3,7 @@ import { SharedDataService } from "src/app/Services/shared-data.service";
 import { CustomerInfoService } from "src/app/Services/CustomerInfo/customer-info.service";
 import { ActivatedRoute } from "@angular/router";
 import { Constant } from "src/app/config/constants";
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-customer-search',
@@ -53,16 +54,29 @@ export class CustomerSearchComponent implements OnInit {
       CustomerTypeID: [null, undefined, ""].includes(this.CustomerTypeID) ? "" : this.CustomerTypeID
     }
   }
+  
   deleteItem(item) {
-    this._sharedDataService.deleteRecord(this.deleteBody(Object.values(item)[0])).subscribe({
-      next: data => {
-        this._sharedDataService.success("Deleted successfully !");
-        this.removeItem(Object.values(item)[0]);
-      },
-      error: error => {
-        this._sharedDataService.error(error)
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._sharedDataService.deleteRecord(this.deleteBody(Object.values(item)[0])).subscribe({
+          next: data => {
+            this._sharedDataService.success("Deleted successfully !");
+            this.removeItem(Object.values(item)[0]);
+          },
+          error: error => {
+            this._sharedDataService.error(error)
+          }
+        });
       }
-    });;
+    });
   }
 
   editItem(item) {
@@ -79,7 +93,7 @@ export class CustomerSearchComponent implements OnInit {
   }
   /* manually remove item from search result */
   removeItem(id) {
-    this.CustomerList = this.CustomerList.filter((searchid) => { Object.values(searchid)[0] != id });
+    this.CustomerList = this.CustomerList.filter((searchid) => { return Object.values(searchid)[0] != id });
   }
 
   /* This will trigger on customer card selection */
